requirements.txt
fastapi~=0.124.4
uvicorn[standard]==0.32.0
sqlmodel==0.0.22
asyncpg==0.30.0
alembic==1.13.3
pydantic-settings==2.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.17
pynacl==1.5.0
sqladmin==0.16.0
slowapi==0.1.9